requirements.txt
certifi==2019.3.9
Django==2.2
django-crispy-forms==1.7.2
Markdown==3.1.1
mistune==0.8.4
Pillow==6.0.0
Pygments==2.4.2
pytz==2019.1
sqlparse==0.3.0
wincertstore==0.2
